python/drf_test/requirements.txt
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
djoser==2.1.0
celery==5.2.7
rabbitmq==0.2.0

